ifFile reader and model maker
main_UPPAAL makes a model struct based on the data of a .if file and
.xtr file.
readIFFile will make a struct of the overall model parameters based on
the data of a .if file. makeStateTransStruct will read in a .xtr file ID
and produces a state struct and transition struct based on the first
state presented.